+++ src/modules/rank.module.ts
@@ -5,9 +5,14 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { Rank } from 'src/entities/rank.entity';
 import { RegisteredGame } from 'src/entities/registered-game.entity';
 import { AuthModule } from 'src/auth/auth.module';
+import { RedisModule } from './redis.module';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Rank, RegisteredGame]), AuthModule],
+  imports: [
+    TypeOrmModule.forFeature([Rank, RegisteredGame]),
+    AuthModule,
+    RedisModule,
+  ],
   exports: [RankService],
   controllers: [RankController],
   providers: [RankService],
